Use consistent rules for filenames of ebuils and misc files.
So far non-ebuild filenames were allowed to contain a : (colon).
This is a problematic character as it is illegal in both Mac OS and
MS-Windows (and there are Prefix profiles for these). Also it has a
special meaning as path separator and as remote host indicator in scp
and other programs.

This limits the allowed characters in filenames to [A-Za-z0-9._+-]
which is the same as IEEE Std 1003.1-2013, section 3.278, with the
addition of the plus character. See also bug 411127.