portage/sync/modules/webrsync: Correct the kwargs.pop() to pop self.spawn_kwargs
With usersync and the repo has portage:portage ownership, the SyncManager.pre_sync()
sets the uid, gid, groups accordingly to spawn the process with.
If the PORTAGE_GPG_DIR is not the same uid, gid, then it will fail to gpg verify.
Dropping these causes the emerge-webrsync binary to run as root, maintaining correct functionality.