portage/sync/modules: Remove the squashdelta module
Due to the original commit of this module being merged untested and with several bugs reported, some not yet fixed, it is
clearly alpha state code.
We, (Alexander, Zac and myself) have agreed that it should be released as a standalone module for further development.
We will consider re-adding it back into mainstream portage when/if it matures, provided the code meets the teams Q/A standards.