require PORTAGE_{BIN,PYM}_PATH not be cleared
Rather than hardcode a full path everywhere as a fallback, assume the
value is always set to the right location.  The current path isn't the
right place anymore already.

If it turns out we want to support this scenario, we can do it via a
bunch of bootstrapping (and symlinked) files.