man/make.conf.5: Document globbing for INSTALL_MASK.
The INSTALL_MASK variable has long supported shell glob patterns, but
these were not mentioned in the man page. In fact there are two
different behaviors -- one for globs that the shell can expand, and
one for "bare globs" like "*.pdf". Both of these are now documented in
the man page.

In addition, a warning and example have been added to the man page
regarding spaces within filenames (patterns).