sync repositories in parallel (bug 557426)
Repos will now be synced in parallel (including their post-sync hooks),
but a given repo will only be synced after its master(s) have synced (in
case that matters for hooks). Output of concurrent processes will be mixed
(irrelevant with --quiet). Support for FEATURES=metadata-transfer will be
handled in the main process, which may be required for some backends (such
as sqlite).

X-Gentoo-Bug: 557426
X-Gentoo-Bug-url: https://bugs.gentoo.org/show_bug.cgi?id=557426
Acked-by: Brian Dolbec <dolsen@gentoo.org>