repoman/repos.py: Fix a regression where the repo is not in repos.conf
The regression was introduced when a variable was used to reduce the size of
some long lines.  The variable was not being reset after the repo was added, so
the remaining code was looking at a stale config.
It turned out the variable was a dupe of an already properly updated class wide
one.

Signed-off-by: Brian Dolbec <dolsen@gentoo.org>