calc_depclean: adjust early warning messages
Suppress "no system list" warning if @profile is non-empty, in order to
support @profile as an alternative to @system. Always warn if the world
file is empty, but drop the "Proceeding is likely to break your
installation" message and 10 second countdown. Abort if a
PackageSetNotFound error is encountered, or if @world is completely empty.
Fix depclean tests to have non-empty @world set.

Acked-by: Alexander Berntsen <bernalex@gentoo.org>