runtests: create a global tempdir to hold subtest files
A lot of unittests currently leak content in /tmp when they run.
Rather than explicitly track down every failing test (which we can
do regardless of this), have the runtest runner create a global
tempdir and use that as a base for children tests.  Then when the
runtest script finishes, it takes care of nuking everything.