repos.conf: support sync-hooks-only-on-change attribute (bug 565172)
If sync-hooks-only-on-change is set to true, do not trigger postsync
hooks unless hooks would have executed for a master repository or the
repository has changed since the previous sync operation.

If the user has not explicitly enabled sync-hooks-only-on-change in
repos.conf, then execute all hooks regardless of whether or not
anything has changed (for backward compatibility).

X-Gentoo-Bug: 565172
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=565172
Acked-by: Brian Dolbec <dolsen@gentoo.org>