egencache: fix results when GIT_DIR is used in the environment.
If GIT_DIR is used, and .git is outside the root of the checkout, then
--work-tree=... needs to be specified, otherwise any Git command that
relies on relative directories to the root will be wrong.

Also ensure that diff-tree's --relative argument is explicitly told
what path to make the relative output for, rather rely on it being
relative to work-tree.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>