Revert "Colorize packages in user sets (bug 577720)"
This reverts commit 739845a6eab95fbc44eab1b7b938c860563631d7.

The feature implemented in this commit for bug 577720 was determined to
be an overcomplicated method of colorizing user sets. Instead, this
feature will be implemented as in bug 583164. The configuration options
implemented in this reverted commit will be removed. Per bug 583164,
packages that belong to a set listed in world_sets will be colorized as
world packages. This eliminates the extra configuration and associated
complications.

See bug 583164 for more details.

Acked-by: Alexander Berntsen <bernalex@gentoo.org>
Acked-by: Zac Medico <zmedico@gentoo.org
Signed-off-by: Alexander Berntsen <bernalex@gentoo.org>