bin/*, misc/emerge-delta-webrsync: migrate to [[.
Right now, there are lots of inconsistencies in the Portage codebase. Some
scripts still use the builtin [ command for performing tests whilst some others
have been migrated to its improved counterpart, the two brackets [[. Sometimes
a script uses a mix of the two which is inconsistent. [[ clearly outnumbers [
in the codebase so we should migrate to the former. This PR attempts to bring
coherence and converts parts still using [ to [[.

See [1] to understand the difference between the two.

Please review.

[1]: http://mywiki.wooledge.org/BashFAQ/031