repoman/actions.py: Add cat/pkg: substtution in the commitsgfile text
Split out the code for msg_prefix().
If the commit message text starts with "cat/pkg: " or "CAT/PKG: ", then it will substitute
that text for the real cat/pkg string deteermined by the msg_prefix().
This allows for a templated commit message that is re-used for several/many packages.