setup.py: enable libc bindings optionally (bug 594744)
The libc bindings are optional, since ctypes is used as a fallback when
they are not available. The libc bindings do not support cross-
compilation, therefore it is useful to be able to build them
conditionally. This patch adds an option to enable them conditionally,
which the ebuild can use by adding the following code to the
python_prepare_all function:

    if use native-extensions; then
        printf "[build_ext]\nportage-ext-modules=true" >> \
            setup.cfg || die
    fi

X-Gentoo-Bug: 594744
X-Gentoo-Bug-URL: https://bugs.gentoo.org/594744
Acked-by: Alexander Berntsen <bernalex@gentoo.org>