emaint: add more meaningful error messages to the logs module
The logs module can fail for a variety of reasons: the PORT_LOGDIR
variable isn't set in make.conf or it doesn't point to a directory; the
PORT_LOGDIR_CLEAN command uses a binary which isn't present in the
system or the binary itself failed during execution. There is only one
generic error message for all these cases. The patch adds error messages
that better describe the reason for the failure.