sync.py: move emerge --sync checks in _sync() and fail when necessary
The emerge --sync repo checks were moved from _get_repos() to _sync() to
make make returning warning and error messages possible.

The existence of the sync-type attribute is now being checked for all
repos, instead of only for the repos given as arguments to emerge --sync.
A warning is printed when a repo without sync-type is found.

A warning is also generated when the repos given to emerge --sync don't
have auto-sync set to 'yes'.

The function _sync() will now fail when there are no valid repos to
sync.

This commit and the commit 'emerge: print messages from
SyncRepos.auto_sync() if --quiet not set' should solve bugs 567478 and
576282.