copyfile: fix compiling against musl libc
The native C module for copying files directly in-kernel is using the
`loff_t` data type, which is defined in fcntl.h. While the file seems to
be indirectly included when building with glibc, it breaks compiling
with musl libc.

Fix the issue by directly including fcntl.h. Define GNU_SOURCE as the
typedef is guarded by this preprocessor macro.