ebuild.sh: Explicitly ban get_libdir in global scope
The value of get_libdir depends on the profile, and so it is not useful
for dependency calculations. Furthermore, it seems that Portage does
not handle defining it in global scope well due to EAPI checking magic.
Ban it completely where it is defined as EAPI function to let developers
catch their mistakes early rather than see them as 'command not found'
errors during dependency calculation / cache updates.

Bug: https://bugs.gentoo.org/629010
Closes: https://github.com/gentoo/portage/pull/197
Reviewed-by: Zac Medico <zmedico@gentoo.org>