Do not enable optimizations by default to work-around pycparser issue
dev-python/pycparser-2.18+ exposes a design flaw in dev-python/ply that
makes it unable to work with -OO code. Given that Gentoo compiles its
Python packages with -OO and both -O and -OO use the same files for
Python < 3.5, using -O in Portage implicitly means -OO for those
implementations.

Remove the optimizations from Portage shebangs to prevent triggering the
issue until we find a proper solution for it.

Bug: https://bugs.gentoo.org/628386
Reviewed-by: Zac Medico <zmedico@gentoo.org>