targetroot: Drop from all EAPIs and remove documentation
Feature flags are frowned upon and this feature has been deemed
unnecessary. It mostly likely would have been used to make packages
depend on themselves but such packages should bootstrap for the build
host within the ebuild instead.

We're not entirely sure whether we want to drop this for good yet so
we'll see how EAPI 7 goes without it before removing all the code.