versions: Drop non-PMS "cvs." prefix in ${PV}
This feature was introduced 12 years ago in (the cvs commit
corresponding to the git commit) 9f3a46665c.  There are a lot
of reasons not to keep it around:

  o PMS permits no such prefix in ${PV}
  o Apparently nobody uses it (perhaps nobody /ever/ used it)
  o It special-cases cvs, which nobody uses, either, in 2017
  o Almost* causes ambiguity between ${PN} and ${PV} in ${P} if
    some future EAPI tried to support both this and dots in
    package names simultaneously.

Therefore, remove it from from the "_v" regular expression,
renumber hard-coded group indexes and nuke corresponding tests.

*PMS would technically avoid abiguity as §3.1.2 requires ${PV}
to "win" any such conflict over contested bits in the middle of
${P}.  However, clearly it's prefereable for this rule to be as
redundant as possible.

Note: this is my 3rd spin of this commitmsg; it's coming along
nicely I think :P

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>