dep_zapdeps: fix virtual/rust handling (bug 645416)
Fix the code from bug 643974 to set the installed_downgrade flag only
if the selected newer package is either installed or in the graph. This
is currently needed for appropriate handling of virtual/rust-1.19.0,
since there's an upgrade to dev-lang/rust-1.23.0 available which may
not be desired since it would mean that dev-lang/rust-bin-1.19.0 has
to be installed in order to satisfy virtual/rust-1.19.0:

 || ( =dev-lang/rust-1.19.0* =dev-lang/rust-bin-1.19.0* )

So, the rust-bin choice is desirable only if the rust-1.23.0 package is
already installed or in the graph.

Fixes: 86ba22da7a2f ("dep_zapdeps: install new package, allow upgrade (bug 643974)")
Bug: https://bugs.gentoo.org/645416